x/toe/types/message_set_pic_test.go
package types

import (
	"testing"

	"github.com/WilliamKelly00/toev2/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSetPic_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSetPic
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSetPic{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSetPic{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
